+++ src/preload.js
@@ -1,4 +1,16 @@
+const { contextBridge } = require('electron')
 const remote = require('@electron/remote');
+
 const mainWindow = remote.getCurrentWindow();
 
-console.log(mainWindow);
+contextBridge.exposeInMainWorld(
+  'TOMATINHO_API',
+  {
+    minimize: () => {
+      mainWindow.minimize();
+    },
+    close: () => {
+      mainWindow.close();
+    },
+  },
+);